Pievienoti koki, salabotas kludas, speleteju markers ari ir, spele ir gatava spelesanai

+++ Assets/Scripts/SceneChanger.cs
@@ -1,15 +1,13 @@
 using System.Collections;
 using System.Collections.Generic;
-using Unity.VisualScripting;
-using UnityEditor.Experimental.GraphView;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 
 public class SceneChanger : MonoBehaviour
 {
-
     public SaveLoadScript saveLoadScript;
     public OverlayScript overlayScript;
+
     public void closeGame()
     {
         StartCoroutine(Delay("quit", -1, ""));
@@ -19,25 +17,41 @@ public class SceneChanger : MonoBehaviour
     {
         if (string.Equals(command, "quit", System.StringComparison.OrdinalIgnoreCase))
         {
-            yield return overlayScript.FadeOut(0.4f);
+            if (overlayScript != null)
+            {
+                yield return overlayScript.FadeOut(0.4f);
+            }
+
             PlayerPrefs.DeleteAll();
 
-    #if UNITY_EDITOR
+#if UNITY_EDITOR
             // Stop play mode in the editor
             UnityEditor.EditorApplication.isPlaying = false;
-    #else
+#else
             Application.Quit();
-    #endif
+#endif
         }
         else if (string.Equals(command, "play", System.StringComparison.OrdinalIgnoreCase))
         {
-            yield return overlayScript.FadeOut(0.4f);
-            saveLoadScript.SaveGame(characterIndex, characterName);
+            if (overlayScript != null)
+            {
+                yield return overlayScript.FadeOut(0.4f);
+            }
+
+            if (saveLoadScript != null)
+            {
+                saveLoadScript.SaveGame(characterIndex, characterName);
+            }
+
             SceneManager.LoadScene(1, LoadSceneMode.Single);
         }
         else if (string.Equals(command, "menu", System.StringComparison.OrdinalIgnoreCase))
         {
-            yield return overlayScript.FadeOut(0.4f);
+            if (overlayScript != null)
+            {
+                yield return overlayScript.FadeOut(0.4f);
+            }
+
             SceneManager.LoadScene(0, LoadSceneMode.Single);
         }
     }
